+++ src/hotspot/share/gc/serial/tenuredGeneration.cpp
@@ -268,14 +268,8 @@ void TenuredGeneration::space_iterate(SpaceClosure* blk,
   blk->do_space(space());
 }
 
-void TenuredGeneration::younger_refs_iterate(OopIterateClosure* blk) {
-  // Apply "cl->do_oop" to (the address of) (exactly) all the ref fields in
-  // "sp" that point into the young generation.
-  // The iteration is only over objects allocated at the start of the
-  // iterations; objects allocated as a result of applying the closure are
-  // not included.
-
-  _rs->younger_refs_in_space_iterate(space(), blk);
+void TenuredGeneration::younger_refs_iterate() {
+  _rs->younger_refs_in_space_iterate(space());
 }
 
 TenuredGeneration::TenuredGeneration(ReservedSpace rs,
